fixes #105
